removing extra renderers and code cleanup. has flow issues (#3439)
* removing extra renderers and code cleanup. has flow issues

* fixing flow errors
