+++ shared/constants/search.js
@@ -80,7 +80,11 @@ export type ToggleUserGroup = TypedAction<'search:toggleUserGroup', {show: boole
 export const reset = 'search:reset'
 export type Reset = TypedAction<'search:reset', {}, void>
 
-export type SearchActions = Search | Results | SelectPlatform | SelectUserForInfo | AddUserToGroup | RemoveUserFromGroup | ToggleUserGroup | Reset
+export const waiting = 'search:waiting'
+export type Waiting = TypedAction<'search:waiting', {waiting: boolean}, void>
+
+export type SearchActions = Search | Results | SelectPlatform | SelectUserForInfo | AddUserToGroup
+  | RemoveUserFromGroup | ToggleUserGroup | Reset | Waiting
 
 export function platformToIcon (platform: SearchPlatforms): IconType {
   return {
