+++ shared/search/dumb.js
@@ -1,7 +1,6 @@
 // @flow
 
-import UserSearch from './user-search/render'
-import UserGroup from './user-search/user-group'
+import Search from './render'
 import type {DumbComponentMap} from '../constants/types/more'
 
 const results = [
@@ -63,85 +62,97 @@ const results = [
   },
 ]
 
-const userSearchMap: DumbComponentMap<UserSearch> = {
-  component: UserSearch,
-  mocks: {
-    'Normal': {
-      onSearch: text => console.log('OnSearch: ', text),
-      searchHintText: 'Search Keybase',
-      searchText: 'malg',
-      searchIcon: 'icon-keybase-logo-32',
-      selectedService: 'Keybase',
-      onClickService: () => console.log('onClickService'),
-      onClickResult: () => console.log('onClickResult'),
-      results,
-      showUserGroup: false,
+const commonUsers = [
+  {
+    service: 'keybase',
+    username: 'max',
+    isFollowing: false,
+    extraInfo: {
+      service: 'none',
+      fullName: 'Max Krohn',
+    },
+  },
+  {
+    service: 'keybase',
+    username: 'malg',
+    isFollowing: false,
+    extraInfo: {
+      service: 'none',
+      fullName: 'John Malg',
+    },
+  },
+  {
+    service: 'external',
+    serviceAvatar: null,
+    serviceName: 'Twitter',
+    icon: 'fa-twitter',
+    username: 'malgorithms',
+    profileUrl: 'https://twitter.com/malgorithms',
+    keybaseSearchResult: null,
+    extraInfo: {
+      service: 'external',
+      icon: 'fa-twitter',
+      serviceUsername: 'malgorithms',
+      serviceAvatar: null,
+      fullNameOnService: 'Chris Coyne',
+    },
+  },
+]
+
+const commonProps = {
+  username: 'bob',
+  onSearch: text => console.log('OnSearch: ', text),
+  searchHintText: 'Search Keybase',
+  searchText: 'malg',
+  searchIcon: 'icon-keybase-logo-32',
+  selectedService: 'Keybase',
+  onClickService: () => console.log('onClickService'),
+  onReset: () => console.log('onReset'),
+  onClickResult: () => console.log('onClickResult'),
+  results,
+  showUserGroup: false,
+  onOpenPrivateGroupFolder: () => console.log('onOpenPrivateGroupFolder'),
+  onOpenPublicGroupFolder: () => console.log('onOpenPublicGroupFolder'),
+  onRemoveUserFromGroup: u => console.log('onRemoveUser', u),
+  onClickUserInGroup: u => console.log('onClickUser', u),
+  onGroupChat: () => console.log('onGroupChat'),
+  onAddUser: () => console.log('onAddUser'),
+  chatEnabled: false,
+  showComingSoon: false,
+  selectedUsers: commonUsers,
+  userForInfoPane: {
+    service: 'keybase',
+    username: 'malg',
+    isFollowing: false,
+    extraInfo: {
+      service: 'none',
+      fullName: 'John Malg',
     },
   },
 }
 
-const userGroupMap: DumbComponentMap<UserGroup> = {
-  component: UserGroup,
+const searchMap: DumbComponentMap<Search> = {
+  component: Search,
   mocks: {
+    'Coming soon': {
+      ...commonProps,
+      showComingSoon: true,
+    },
+    'Searching': {
+      ...commonProps,
+    },
+    'Group': {
+      ...commonProps,
+      showUserGroup: true,
+    },
     'Chat enabled': {
-      parentProps: {style: {marginTop: 20, flex: 1}},
-      userForInfoPane: {
-        service: 'keybase',
-        username: 'malg',
-        isFollowing: false,
-        extraInfo: {
-          service: 'none',
-          fullName: 'John Malg',
-        },
-      },
-      users: [
-        {
-          service: 'keybase',
-          username: 'max',
-          isFollowing: false,
-          extraInfo: {
-            service: 'none',
-            fullName: 'Max Krohn',
-          },
-        },
-        {
-          service: 'keybase',
-          username: 'malg',
-          isFollowing: false,
-          extraInfo: {
-            service: 'none',
-            fullName: 'John Malg',
-          },
-        },
-        {
-          service: 'external',
-          serviceAvatar: null,
-          serviceName: 'Twitter',
-          icon: 'fa-twitter',
-          username: 'malgorithms',
-          profileUrl: 'https://twitter.com/malgorithms',
-          keybaseSearchResult: null,
-          extraInfo: {
-            service: 'external',
-            icon: 'fa-twitter',
-            serviceUsername: 'malgorithms',
-            serviceAvatar: null,
-            fullNameOnService: 'Chris Coyne',
-          },
-        },
-      ],
-      onOpenPrivateGroupFolder: () => console.log('onOpenPrivateGroupFolder'),
-      onOpenPublicGroupFolder: () => console.log('onOpenPublicGroupFolder'),
-      onRemoveUser: u => console.log('onRemoveUser', u),
-      onClickUser: u => console.log('onClickUser', u),
-      onGroupChat: () => console.log('onGroupChat'),
-      onAddUser: () => console.log('onAddUser'),
+      ...commonProps,
+      showUserGroup: true,
       chatEnabled: true,
     },
   },
 }
 
 export default {
-  'user search': userSearchMap,
-  'user group': userGroupMap,
+  'Search': searchMap,
 }
