+++ shared/search/render.desktop.js
@@ -3,39 +3,25 @@ import React, {Component} from 'react'
 import {Box, Text} from '../common-adapters'
 import SearchHelp from './help.desktop'
 import {globalStyles, globalColors} from '../styles/style-guide'
-import {SearchContainer, SearchBar, searchResultsList} from './user-search/render.desktop'
+import UserSearch from './user-search/render.desktop'
 import UserGroup from './user-search/user-group'
+import SearchBar from './user-search/search-bar.desktop'
 import UserPane from './user-pane'
 
 import type {Props} from './render'
+import type {Props as UserSearchProps} from './user-search/render'
+import type {Props as UserGroupProps} from './user-search/user-group'
+import type {Props as SearchBarProps} from './user-search/search-bar'
+
+const ClearSearch = ({onReset, showUserGroup}) => (
+  <Box style={{...globalStyles.flexBoxColumn, height: 48, justifyContent: 'center', alignItems: 'flex-end', paddingRight: 16}}>
+    {showUserGroup && <Text type='BodySmallSecondaryLink' onClick={onReset}>Clear search</Text>}
+  </Box>
+)
 
 class Render extends Component<void, Props, void> {
   _renderComingSoon () {
     return <SearchHelp username={this.props.username} />
-  }
-
-  _renderInfoPane () {
-    return (
-      <Box style={{boxShadow: `0 0 5px ${globalColors.black_20}`, overflow: 'hidden'}}>
-        <UserPane />
-      </Box>
-    )
-  }
-
-  _renderSearchResultsOrGroupAdd () {
-    return (
-      <Box style={{overflowY: 'auto', height: 'calc(100% - 96px)'}}>
-        {this.props.showUserGroup ? <UserGroup
-          users={this.props.selectedUsers}
-          userForInfoPane={this.props.userForInfoPane}
-          onRemoveUser={this.props.onRemoveUserFromGroup}
-          onClickUser={this.props.onClickUserInGroup}
-          onOpenPrivateGroupFolder={this.props.onOpenPrivateGroupFolder}
-          onOpenPublicGroupFolder={this.props.onOpenPublicGroupFolder}
-          onGroupChat={this.props.onGroupChat}
-          chatEnabled={this.props.chatEnabled} /> : searchResultsList(this.props)}
-      </Box>
-    )
   }
 
   render () {
@@ -43,21 +29,22 @@ class Render extends Component<void, Props, void> {
       return this._renderComingSoon()
     }
 
+    const userSearchProps: UserSearchProps = this.props
+    const userGroupProps: UserGroupProps = this.props
+    const searchBarProps: SearchBarProps = this.props
+
     return (
       <Box style={{...globalStyles.flexBoxRow, flex: 1}}>
-        <SearchContainer>
-          <Box style={{...globalStyles.flexBoxColumn, height: 48, justifyContent: 'center', alignItems: 'flex-end', paddingRight: 16}}>
-            {this.props.showUserGroup && <Text type='BodySmallSecondaryLink' onClick={this.props.onReset}>Clear search</Text>}
+        <Box style={{...globalStyles.flexBoxColumn, flex: 1}}>
+          <ClearSearch {...this.props} />
+          <SearchBar {...searchBarProps} />
+          <Box style={{overflowY: 'auto', height: 'calc(100% - 96px)'}}>
+            {this.props.showUserGroup ? <UserGroup {...userGroupProps} /> : <UserSearch {...userSearchProps} />}
           </Box>
-          <SearchBar
-            onClickService={this.props.onClickService}
-            onSearch={this.props.onSearch}
-            searchHintText={this.props.searchHintText}
-            searchText={this.props.searchText}
-            selectedService={this.props.selectedService} />
-          {this._renderSearchResultsOrGroupAdd()}
-        </SearchContainer>
-        {this._renderInfoPane()}
+        </Box>
+        <Box style={{boxShadow: `0 0 5px ${globalColors.black_20}`, overflow: 'hidden'}}>
+          <UserPane />
+        </Box>
       </Box>
     )
   }
