+++ shared/search/user-search/render.desktop.js
@@ -1,14 +1,11 @@
 // @flow
 import React, {Component} from 'react'
-import _ from 'lodash'
 
-import {Avatar, Box, ClickableBox, Icon, Input, Text} from '../../common-adapters'
-import {globalStyles, globalColors, transition} from '../../styles/style-guide'
-import {platformToLogo24} from '../../constants/search'
+import {Avatar, Box, ClickableBox, Icon, Text} from '../../common-adapters'
+import {globalStyles, globalColors} from '../../styles/style-guide'
 
-import type {SearchResult, SearchPlatforms} from '../../constants/search'
-import type {Props, SearchResultFn, ServiceFn} from './render'
-import type {IconType} from '../../common-adapters/icon'
+import type {SearchResult} from '../../constants/search'
+import type {Props, SearchResultFn} from './render'
 import type {Props as TextProps} from '../../common-adapters/text'
 
 function EmboldenTextMatch ({text, match, style, textType, emboldenStyle}: {text: string, match: string, emboldenStyle?: Object, style?: Object, textType: TextProps.type}) {
@@ -120,185 +117,18 @@ export function Result ({result, searchText, onClickResult}: {result: SearchResu
   )
 }
 
-type ServiceIconState = {showingTooltip: boolean}
-type ServiceIconProps = {serviceName: SearchPlatforms, tooltip: string,
-  iconType: IconType, selected: boolean, onClickService: ServiceFn}
-class ServiceIcon extends Component<void, ServiceIconProps, ServiceIconState> {
-  state: ServiceIconState;
-
-  constructor (props: ServiceIconProps) {
-    super(props)
-
-    this.state = {
-      showingTooltip: false,
-    }
-  }
-
+class Render extends Component<void, Props, void> {
   render () {
-    const {serviceName, tooltip, iconType, selected, onClickService} = this.props
     return (
-      <Box style={{...serviceContainerStyle, backgroundColor: selected ? globalColors.blue4 : null}}
-        onMouseEnter={() => this.setState({showingTooltip: true})}
-        onMouseLeave={() => this.setState({showingTooltip: false})}
-        onClick={() => onClickService(serviceName)} >
-        <Icon type={iconType} style={{...serviceIconStyle,
-          opacity: selected || this.state.showingTooltip ? 1.0 : 0.6}} />
-        <Text type='BodyXSmall' style={{...serviceTooltipStyle,
-          opacity: this.state.showingTooltip ? 1 : 0}}>{tooltip}</Text>
+      <Box style={{overflowY: 'auto', flex: 1}}>
+        {this.props.waiting && <Icon type='icon-progress-grey-animated'
+          style={{position: 'absolute', top: 20, left: 0, right: 0, marginLeft: 'auto', marginRight: 'auto'}} />}
+        {this.props.results.map(r => (
+          <Result key={r.service + (r.icon ? r.icon : '') + r.username} result={r}
+            searchText={this.props.searchText || ''} onClickResult={this.props.onClickResult} />))}
       </Box>
     )
   }
 }
 
-export type SearchBarProps = {
-  selectedService: ?SearchPlatforms,
-  onSearch: (term: string, platform?: ?SearchPlatforms) => void,
-  searchText: ?string,
-  searchHintText: string,
-  onClickService: (service: SearchPlatforms) => void,
-}
-
-export class SearchBar extends Component<void, SearchBarProps, void> {
-  _onDebouncedSearch: (overridePlatform?: SearchPlatforms) => void;
-
-  constructor (props: SearchBarProps) {
-    super(props)
-    this._onDebouncedSearch = _.debounce(this._onSearch, 500)
-  }
-
-  componentWillReceiveProps (nextProps: SearchBarProps) {
-    if (nextProps.searchText === null && nextProps.searchText !== this.props.searchText) {
-      this.refs && this.refs.searchBox && this.refs.searchBox.clearValue()
-    }
-  }
-
-  _onSearch (overridePlatform?: SearchPlatforms) {
-    this.props.onSearch(this.refs.searchBox ? this.refs.searchBox.getValue() : '', overridePlatform || null)
-  }
-
-  _onClickService (platform: SearchPlatforms) {
-    this.props.onClickService(platform)
-    if (this.refs.searchBox) {
-      if (this.refs.searchBox.getValue()) {
-        this._onSearch(platform)
-      }
-      this.refs.searchBox.focus()
-    }
-  }
-
-  render () {
-    const services = ['Keybase', 'Twitter', 'Github', 'Coinbase', 'Reddit', 'Hackernews']
-    const tooltips: {[key: string]: ?string} = {'Hackernews': 'Hacker News'}
-
-    return (
-      <Box>
-        <Box style={stylesServicesContainer}>
-          {services.map(s => (
-            <ServiceIcon
-              key={s}
-              serviceName={s}
-              tooltip={tooltips[s] || s}
-              iconType={platformToLogo24(s)}
-              selected={this.props.selectedService === s}
-              onClickService={p => this._onClickService(p)}
-              />
-          ))}
-        </Box>
-        <Box style={stylesInputContainer}>
-          <Input
-            type='text'
-            autoFocus
-            ref='searchBox'
-            onEnterKeyDown={() => this._onSearch()}
-            onChange={() => this._onDebouncedSearch()}
-            value={this.props.searchText}
-            hintText={this.props.searchHintText}
-            hintStyle={{textAlign: 'left', marginTop: 3}}
-            underlineShow={false}
-            style={stylesInput}
-            textStyle={{height: 31}} />
-          <Icon type='fa-kb-iconfont-remove' style={{marginRight: 16, opacity: this.props.searchText ? 1 : 0}}
-            onClick={() => this.refs.searchBox.clearValue()} />
-        </Box>
-      </Box>
-    )
-  }
-}
-
-export function searchResultsList ({results, searchText, onClickResult}: {results: Array<SearchResult>, searchText: ?string, onClickResult: SearchResultFn}) {
-  return results.map(r => (
-    <Result key={r.service + (r.icon ? r.icon : '') + r.username} result={r} searchText={searchText || ''} onClickResult={onClickResult} />
-  ))
-}
-
-export class SearchContainer extends Component {
-  render () {
-    return (
-      <Box style={stylesContainer}>
-        {this.props.children}
-      </Box>
-    )
-  }
-}
-
-export default class Render extends Component<void, Props, void> {
-  render () {
-    return (
-      <SearchContainer>
-        <SearchBar {...this.props} />
-        <Box style={{overflowY: 'auto', flex: 1}}>
-          {searchResultsList(this.props)}
-        </Box>
-      </SearchContainer>
-    )
-  }
-}
-
-const stylesContainer = {
-  ...globalStyles.flexBoxColumn,
-  flex: 1,
-}
-const stylesServicesContainer = {
-  ...globalStyles.flexBoxRow,
-  height: 48,
-  paddingLeft: 16,
-}
-const stylesInputContainer = {
-  ...globalStyles.flexBoxRow,
-  height: 48,
-  borderBottom: `solid 1px ${globalColors.black_10}`,
-  alignItems: 'center',
-  marginBottom: 8,
-}
-const stylesInput = {
-  flex: 1,
-  textAlign: 'left',
-  marginLeft: 16,
-  marginRight: 30,
-}
-const serviceContainerStyle = {
-  ...globalStyles.flexBoxColumn,
-  ...transition('backgroundColor'),
-  alignItems: 'center',
-  borderRadius: 25,
-  height: 50,
-  justifyContent: 'center',
-  position: 'relative',
-  width: 50,
-}
-const serviceIconStyle = {
-  ...transition('opacity'),
-}
-const serviceTooltipStyle = {
-  ...transition('opacity'),
-  backgroundColor: globalColors.black_40,
-  borderRadius: 65,
-  color: globalColors.white,
-  left: -16,
-  lineHeight: '22px',
-  minHeight: 22,
-  minWidth: 86,
-  position: 'absolute',
-  textAlign: 'center',
-  top: -24,
-}
+export default Render
