+++ shared/search/user-search/search-bar.js.flow
@@ -1,0 +1,13 @@
+/* @flow */
+import React, {Component} from 'react'
+import type {SearchResult, SearchPlatforms} from '../../constants/search'
+
+export type Props = {
+  selectedService: ?SearchPlatforms,
+  onSearch: (term: string, platform?: ?SearchPlatforms) => void,
+  searchText: ?string,
+  searchHintText: string,
+  onClickService: (service: SearchPlatforms) => void,
+}
+
+export class SearchBar extends Component<void, Props, void> {}
